add name and accordions
