+++ packages/compass-data-modeling/src/services/data-model-storage.ts
@@ -11,6 +11,7 @@ export type RelationshipSide = z.output<typeof RelationshipSideSchema>;
 
 export const RelationshipSchema = z.object({
   id: z.string().uuid(),
+  name: z.string().optional(),
   relationship: z.tuple([RelationshipSideSchema, RelationshipSideSchema]),
   isInferred: z.boolean(),
 });
